app/requirements.txt
fastapi
uvicorn
streamlit
torch
pillow
requests
python-multipart
mlflow